+++ tasty-plutus/CHANGELOG.md
@@ -1,5 +1,17 @@
-# Revision history for plutus-extra
+# Revision history for `tasty-plutus`
 
-## 1.0
+This format is based on [Keep A Changelog](https://keepachangelog.com/en/1.0.0)
 
-- Initial release.
+## Unreleased
+
+### Changed
+
+* Better rendering of record-like things. 
+
+### Removed
+
+* All `Pretty` instances.
+
+## 1.0 -- 2021-09-08
+
+* Initial release.